+++ ChiTech/ChiMath/SpatialDiscretization/FiniteElement/PiecewiseLinear/pwlc.h
@@ -95,25 +95,25 @@ private:
 
 public:
   //03
-  void BuildSparsityPattern(std::vector<int>& nodal_nnz_in_diag,
-                            std::vector<int>& nodal_nnz_off_diag,
+  void BuildSparsityPattern(std::vector<int64_t>& nodal_nnz_in_diag,
+                            std::vector<int64_t>& nodal_nnz_off_diag,
                             chi_math::UnknownManager& unknown_manager) override;
 
   //04 Mappings
-  int MapDOF(const chi_mesh::Cell& cell,
-             unsigned int node,
-             const chi_math::UnknownManager& unknown_manager,
-             unsigned int unknown_id,
-             unsigned int component) const override;
-  int MapDOFLocal(const chi_mesh::Cell& cell,
-                  unsigned int node,
-                  const chi_math::UnknownManager& unknown_manager,
-                  unsigned int unknown_id,
-                  unsigned int component) const override;
+  int64_t MapDOF(const chi_mesh::Cell& cell,
+                 unsigned int node,
+                 const chi_math::UnknownManager& unknown_manager,
+                 unsigned int unknown_id,
+                 unsigned int component) const override;
+  int64_t MapDOFLocal(const chi_mesh::Cell& cell,
+                      unsigned int node,
+                      const chi_math::UnknownManager& unknown_manager,
+                      unsigned int unknown_id,
+                      unsigned int component) const override;
 
-  int MapDOF(const chi_mesh::Cell& cell, unsigned int node) const override
+  int64_t MapDOF(const chi_mesh::Cell& cell, unsigned int node) const override
   { return MapDOF(cell,node,ChiMath::UNITARY_UNKNOWN_MANAGER,0,0); }
-  int MapDOFLocal(const chi_mesh::Cell& cell, unsigned int node) const override
+  int64_t MapDOFLocal(const chi_mesh::Cell& cell, unsigned int node) const override
   { return MapDOFLocal(cell,node,ChiMath::UNITARY_UNKNOWN_MANAGER,0,0); }
 
   //05
