+++ ChiTech/ChiMath/SpatialDiscretization/FiniteVolume/fv_03_buildsparsity.cc
@@ -6,8 +6,8 @@
 //###################################################################
 /**Builds finite volume based sparsity pattern.*/
 void SpatialDiscretization_FV::BuildSparsityPattern(
-  std::vector<int> &nodal_nnz_in_diag,
-  std::vector<int> &nodal_nnz_off_diag,
+  std::vector<int64_t> &nodal_nnz_in_diag,
+  std::vector<int64_t> &nodal_nnz_off_diag,
   chi_math::UnknownManager& unknown_manager)
 {
   unsigned int num_uk = unknown_manager.unknowns.size(); //Number of unknowns
